Adds extra paramaters to the getinfo rpc call.
Adds in the unconfirmed and immature parameters to the getinfo rpc call
which are shown in the QT wallet but not available via the cli by default.

unconfirmed shows coins that have been received but have yet to be fully
confirmed.
immature shows coins temporarily used for a Proof of Stake reward and will
return to main balance once the PoS block has become confirmed